Move AccountInfo and related classes to their own directory
