+++ b2/account_info/__init__.py
@@ -1,0 +1,5 @@
+from .sqlite_account_info import SqliteAccountInfo
+from .upload_url_pool import UploadUrlPool
+
+assert SqliteAccountInfo
+assert UploadUrlPool
