+++ b2/account_info/abstract.py
@@ -1,0 +1,136 @@
+######################################################################
+#
+# File: b2/account_info/abstract.py
+#
+# Copyright 2016 Backblaze Inc. All Rights Reserved.
+#
+# License https://www.backblaze.com/using_b2_code.html
+#
+######################################################################
+
+
+from abc import (ABCMeta, abstractmethod)
+
+import six
+
+
+@six.add_metaclass(ABCMeta)
+class AbstractAccountInfo(object):
+    """
+    Holder for all account-related information that needs to be kept
+    between API calls, and between invocations of the command-line
+    tool.  This includes: account id, application key, auth tokens,
+    API URL, download URL, and uploads URLs.
+
+    This class must be THREAD SAFE because it may be used by multiple
+    threads running in the same Python process.  It also needs to be
+    safe against multiple processes running at the same time.
+    """
+
+    REALM_URLS = {
+        'production': 'https://api.backblaze.com',
+        'dev': 'http://api.test.blaze:8180',
+        'staging': 'https://api.backblaze.net',
+    }
+
+    @abstractmethod
+    def clear(self):
+        """
+        Removes all stored information
+        """
+
+    @abstractmethod
+    def refresh_entire_bucket_name_cache(self, name_id_iterable):
+        """
+        Removes all previous name-to-id mappings and stores new ones.
+        """
+
+    @abstractmethod
+    def remove_bucket_name(self, bucket_name):
+        """
+        Removes one entry from the bucket name cache.
+        """
+
+    @abstractmethod
+    def save_bucket(self, bucket):
+        """
+        Remembers the ID for a bucket name.
+        """
+
+    @abstractmethod
+    def get_bucket_id_or_none_from_bucket_name(self, bucket_name):
+        """
+        Looks up the bucket ID for a given bucket name.
+        """
+
+    @abstractmethod
+    def clear_bucket_upload_data(self, bucket_id):
+        """
+        Removes all upload URLs for the given bucket.
+        """
+
+    @abstractmethod
+    def get_account_id(self):
+        """ returns account_id or raises MissingAccountData exception """
+
+    @abstractmethod
+    def get_account_auth_token(self):
+        """ returns account_auth_token or raises MissingAccountData exception """
+
+    @abstractmethod
+    def get_api_url(self):
+        """ returns api_url or raises MissingAccountData exception """
+
+    @abstractmethod
+    def get_application_key(self):
+        """ returns application_key or raises MissingAccountData exception """
+
+    @abstractmethod
+    def get_download_url(self):
+        """ returns download_url or raises MissingAccountData exception """
+
+    @abstractmethod
+    def get_realm(self):
+        """ returns realm or raises MissingAccountData exception """
+
+    @abstractmethod
+    def get_minimum_part_size(self):
+        """
+        :return: returns the minimum number of bytes in a part of a large file
+        """
+
+    @abstractmethod
+    def set_auth_data(
+        self, account_id, auth_token, api_url, download_url, minimum_part_size, application_key,
+        realm
+    ):
+        pass
+
+    @abstractmethod
+    def take_bucket_upload_url(self, bucket_id):
+        """
+        Returns a pair (upload_url, upload_auth_token) that has been removed
+        from the pool for this bucket, or (None, None) if there are no more
+        left.
+        """
+
+    @abstractmethod
+    def put_bucket_upload_url(self, bucket_id, upload_url, upload_auth_token):
+        """
+        Add an (upload_url, upload_auth_token) pair to the pool available for
+        the bucket.
+        """
+
+    @abstractmethod
+    def put_large_file_upload_url(self, file_id, upload_url, upload_auth_token):
+        pass
+
+    @abstractmethod
+    def take_large_file_upload_url(self, file_id):
+        pass
+
+    @abstractmethod
+    def clear_large_file_upload_urls(self, file_id):
+        pass
+
+
