+++ b2/account_info/sqlite_account_info.py
@@ -1,6 +1,6 @@
 ######################################################################
 #
-# File: b2/account_info.py
+# File: b2/account_info/sqlite_account_info.py
 #
 # Copyright 2016 Backblaze Inc. All Rights Reserved.
 #
@@ -8,173 +8,15 @@
 #
 ######################################################################
 
-import collections
 import json
 import os
 import sqlite3
 import stat
 import threading
-from abc import (ABCMeta, abstractmethod)
 
-import six
-
+from .abstract import AbstractAccountInfo
 from .exception import (CorruptAccountInfo, MissingAccountData)
-
-
-@six.add_metaclass(ABCMeta)
-class AbstractAccountInfo(object):
-    """
-    Holder for all account-related information that needs to be kept
-    between API calls, and between invocations of the command-line
-    tool.  This includes: account id, application key, auth tokens,
-    API URL, download URL, and uploads URLs.
-
-    This class must be THREAD SAFE because it may be used by multiple
-    threads running in the same Python process.  It also needs to be
-    safe against multiple processes running at the same time.
-    """
-
-    REALM_URLS = {
-        'production': 'https://api.backblaze.com',
-        'dev': 'http://api.test.blaze:8180',
-        'staging': 'https://api.backblaze.net',
-    }
-
-    @abstractmethod
-    def clear(self):
-        """
-        Removes all stored information
-        """
-
-    @abstractmethod
-    def refresh_entire_bucket_name_cache(self, name_id_iterable):
-        """
-        Removes all previous name-to-id mappings and stores new ones.
-        """
-
-    @abstractmethod
-    def remove_bucket_name(self, bucket_name):
-        """
-        Removes one entry from the bucket name cache.
-        """
-
-    @abstractmethod
-    def save_bucket(self, bucket):
-        """
-        Remembers the ID for a bucket name.
-        """
-
-    @abstractmethod
-    def get_bucket_id_or_none_from_bucket_name(self, bucket_name):
-        """
-        Looks up the bucket ID for a given bucket name.
-        """
-
-    @abstractmethod
-    def clear_bucket_upload_data(self, bucket_id):
-        """
-        Removes all upload URLs for the given bucket.
-        """
-
-    @abstractmethod
-    def get_account_id(self):
-        """ returns account_id or raises MissingAccountData exception """
-
-    @abstractmethod
-    def get_account_auth_token(self):
-        """ returns account_auth_token or raises MissingAccountData exception """
-
-    @abstractmethod
-    def get_api_url(self):
-        """ returns api_url or raises MissingAccountData exception """
-
-    @abstractmethod
-    def get_application_key(self):
-        """ returns application_key or raises MissingAccountData exception """
-
-    @abstractmethod
-    def get_download_url(self):
-        """ returns download_url or raises MissingAccountData exception """
-
-    @abstractmethod
-    def get_realm(self):
-        """ returns realm or raises MissingAccountData exception """
-
-    @abstractmethod
-    def get_minimum_part_size(self):
-        """
-        :return: returns the minimum number of bytes in a part of a large file
-        """
-
-    @abstractmethod
-    def set_auth_data(
-        self, account_id, auth_token, api_url, download_url, minimum_part_size, application_key,
-        realm
-    ):
-        pass
-
-    @abstractmethod
-    def take_bucket_upload_url(self, bucket_id):
-        """
-        Returns a pair (upload_url, upload_auth_token) that has been removed
-        from the pool for this bucket, or (None, None) if there are no more
-        left.
-        """
-
-    @abstractmethod
-    def put_bucket_upload_url(self, bucket_id, upload_url, upload_auth_token):
-        """
-        Add an (upload_url, upload_auth_token) pair to the pool available for
-        the bucket.
-        """
-
-    @abstractmethod
-    def put_large_file_upload_url(self, file_id, upload_url, upload_auth_token):
-        pass
-
-    @abstractmethod
-    def take_large_file_upload_url(self, file_id):
-        pass
-
-    @abstractmethod
-    def clear_large_file_upload_urls(self, file_id):
-        pass
-
-
-class UploadUrlPool(object):
-    """
-    For each key (either a bucket id or large file id), holds a pool
-    of (url, auth_token) pairs, with thread-safe methods to add and
-    remove them.
-    """
-
-    def __init__(self):
-        self._lock = threading.Lock()
-        self._pool = collections.defaultdict(list)
-
-    def put(self, key, url, auth_token):
-        """
-        Adds the url and auth token to the pool for the given key.
-        """
-        with self._lock:
-            pair = (url, auth_token)
-            self._pool[key].append(pair)
-
-    def take(self, key):
-        """
-        Returns (url, auth_token) if one is available, or (None, None) if not.
-        """
-        with self._lock:
-            pair_list = self._pool[key]
-            if pair_list:
-                return pair_list.pop()
-            else:
-                return (None, None)
-
-    def clear_for_key(self, key):
-        with self._lock:
-            if key in self._pool:
-                del self._pool[key]
+from .upload_url_pool import (UploadUrlPool)
 
 
 class SqliteAccountInfo(AbstractAccountInfo):
@@ -421,140 +263,3 @@ class SqliteAccountInfo(AbstractAccountInfo):
 
     def clear_large_file_upload_urls(self, file_id):
         self._large_file_uploads.clear_for_key(file_id)
-
-
-class StubAccountInfo(AbstractAccountInfo):
-
-    REALM_URLS = {'production': 'http://production.example.com'}
-
-    def __init__(self):
-        self.clear()
-
-    def clear(self):
-        self.account_id = None
-        self.auth_token = None
-        self.api_url = None
-        self.download_url = None
-        self.minimum_part_size = None
-        self.application_key = None
-        self.realm = None
-        self.buckets = {}
-        self._large_file_uploads = collections.defaultdict(list)
-
-    def clear_bucket_upload_data(self, bucket_id):
-        if bucket_id in self.buckets:
-            del self.buckets[bucket_id]
-
-    def set_auth_data(
-        self, account_id, auth_token, api_url, download_url, minimum_part_size, application_key,
-        realm
-    ):
-        self.account_id = account_id
-        self.auth_token = auth_token
-        self.api_url = api_url
-        self.download_url = download_url
-        self.minimum_part_size = minimum_part_size
-        self.application_key = application_key
-        self.realm = realm
-
-    def refresh_entire_bucket_name_cache(self, name_id_iterable):
-        self.buckets = {}
-
-    def get_bucket_id_or_none_from_bucket_name(self, bucket_name):
-        return None
-
-    def save_bucket(self, bucket):
-        pass
-
-    def remove_bucket_name(self, bucket_name):
-        pass
-
-    def take_bucket_upload_url(self, bucket_id):
-        return (None, None)
-
-    def put_bucket_upload_url(self, bucket_id, upload_url, upload_auth_token):
-        pass
-
-    def get_account_id(self):
-        return self.account_id
-
-    def get_account_auth_token(self):
-        return self.auth_token
-
-    def get_api_url(self):
-        return self.api_url
-
-    def get_application_key(self):
-        return self.application_key
-
-    def get_download_url(self):
-        return self.download_url
-
-    def get_minimum_part_size(self):
-        return self.minimum_part_size
-
-    def get_realm(self):
-        return self.realm
-
-    def get_bucket_upload_data(self, bucket_id):
-        return self.buckets.get(bucket_id, (None, None))
-
-    def put_large_file_upload_url(self, file_id, upload_url, upload_auth_token):
-        self._large_file_uploads[file_id].append((upload_url, upload_auth_token))
-
-    def take_large_file_upload_url(self, file_id):
-        upload_urls = self._large_file_uploads.get(file_id, [])
-        if len(upload_urls) == 0:
-            return (None, None)
-        else:
-            return upload_urls.pop()
-
-    def clear_large_file_upload_urls(self, file_id):
-        if file_id in self._large_file_uploads:
-            del self._large_file_uploads[file_id]
-
-
-def test_upload_url_concurrency():
-    # Clean up from previous tests
-    file_name = '/tmp/test_upload_conncurrency.db'
-    try:
-        os.unlink(file_name)
-    except OSError:
-        pass
-
-    # Make an account info with a bunch of upload URLs in it.
-    account_info = SqliteAccountInfo(file_name)
-    available_urls = set()
-    for i in six.moves.range(3000):
-        url = 'url_%d' % i
-        account_info.put_bucket_upload_url('bucket-id', url, 'auth-token-%d' % i)
-        available_urls.add(url)
-
-    # Pull them all from the account info, from multiple threads
-    lock = threading.Lock()
-
-    def run_thread():
-        while True:
-            (url, _) = account_info.take_bucket_upload_url('bucket-id')
-            if url is None:
-                break
-            with lock:
-                if url in available_urls:
-                    available_urls.remove(url)
-                else:
-                    print('DOUBLE:', url)
-
-    threads = []
-    for i in six.moves.range(5):
-        thread = threading.Thread(target=run_thread)
-        thread.start()
-        threads.append(thread)
-    for t in threads:
-        t.join()
-
-    # Check
-    if len(available_urls) != 0:
-        print('LEAK:', available_urls)
-
-    # Clean up
-    os.unlink(file_name)
