+++ b2/account_info/upload_url_pool.py
@@ -1,3 +1,14 @@
+######################################################################
+#
+# File: b2/account_info/upload_url_pool.py
+#
+# Copyright 2016 Backblaze Inc. All Rights Reserved.
+#
+# License https://www.backblaze.com/using_b2_code.html
+#
+######################################################################
+
+
 import collections
 import threading
