+++ b2/api.py
@@ -9,11 +9,12 @@
 ######################################################################
 
 from .account_info import SqliteAccountInfo
+from .account_info.exception import MissingAccountData
 from .b2http import B2Http
 from .bucket import Bucket, BucketFactory
 from .cache import AuthInfoCache, DummyCache
 from .download_dest import DownloadDestProgressWrapper
-from .exception import MissingAccountData, NonExistentBucket
+from .exception import NonExistentBucket
 from .file_version import FileVersionInfoFactory, FileIdAndName
 from .part import PartFactory
 from .progress import DoNothingProgressListener
