+++ b2/console_tool.py
@@ -20,11 +20,12 @@ import textwrap
 import six
 
 from .account_info import (SqliteAccountInfo, test_upload_url_concurrency)
+from .account_info.exception import (MissingAccountData)
 from .api import (B2Api)
 from .b2http import (test_http)
 from .cache import (AuthInfoCache)
 from .download_dest import (DownloadDestLocalFile)
-from .exception import (B2Error, BadFileInfo, MissingAccountData)
+from .exception import (B2Error, BadFileInfo)
 from .file_version import (FileVersionInfo)
 from .parse_args import parse_arg_list
 from .progress import (make_progress_listener)
