+++ test/test_account_info.py
@@ -16,8 +16,8 @@ import unittest
 
 import six
 
-from b2.upload_url_pool import UploadUrlPool
-from b2.exception import CorruptAccountInfo, MissingAccountData
+from b2.account_info.upload_url_pool import UploadUrlPool
+from b2.account_info.exception import CorruptAccountInfo, MissingAccountData
 
 if not platform.system().lower().startswith('java'):
     # in Jython 2.7.1b3 there is no sqlite3
