+++ test/test_bucket.py
@@ -16,7 +16,7 @@ import unittest
 
 import six
 
-from b2.account_info import StubAccountInfo
+from stub_account_info import StubAccountInfo
 from b2.api import B2Api
 from b2.bucket import LargeFileUploadState
 from b2.download_dest import DownloadDestBytes
